Show raw text from events
