+++ apps/arcade-city/src/views/RequestFeed.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { FlatList, StyleSheet, View } from 'react-native'
-import { RideRequest as RideRequestType } from '../lib/nostr/store'
+import { RideRequest as RideRequestType } from '../lib/nostr/types'
 import { RideRequest } from './RideRequest'
 
 export const RequestFeed = ({ requests }: { requests: RideRequestType[] }) => {
@@ -10,7 +10,6 @@ export const RequestFeed = ({ requests }: { requests: RideRequestType[] }) => {
     })
     .slice(20)
   const key = 'id'
-
   const arrayUniqueByKey = [...new Map(sortedRequests.map((item) => [item[key], item])).values()]
   return (
     <View style={styles.container}>
